refactor: reorganize package to standardize with keyrus ds packaging
